backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Gemini AI
google-genai>=1.0.0

# PDF Processing
pymupdf>=1.23.0
pdf2image>=1.16.3

# Image Processing
Pillow>=10.0.0

# CORS
python-dotenv>=1.0.0
